Scored Baskets Reset Positions
Decided to implement a component node I'm calling the "gamemode core" which will be responsible for the majority of the game's logic around scores, teams, timers, effects, etc. This component is functioning and able to reset the positions of the game's objects (inefficiently, probably, but I want this thing DONE not perfect or even great lmao).